Likelihood for SecSSE model

Logikelihood calculation for the SecSSE model given a set of parameters and data, returning also the likelihoods along the branches

Usage

secsse_loglik_eval(
  parameter,
  phy,
  traits,
  num_concealed_states,
  ancestral_states,
  cond = "proper_cond",
  root_state_weight = "proper_weights",
  sampling_fraction,
  setting_calculation = NULL,
  loglik_penalty = 0,
  is_complete_tree = FALSE,
  atol = 1e-12,
  rtol = 1e-12,
  method = "odeint::bulirsch_stoer",
  num_steps = NULL,
  verbose = FALSE
)

Arguments

parameter

list where first vector represents lambdas, the second mus and the third transition rates.

phy

phylogenetic tree of class phylo, ultrametric, fully-resolved, rooted and with branch lengths.

traits

vector with trait states, order of states must be the same as tree tips, for help, see vignette.

num_concealed_states

number of concealed states, generally equivalent to number of examined states.

ancestral_states

ancestral states matrix provided by secsse_loglik, this is used as starting points for the branch integration

cond

condition on the existence of a node root: "maddison_cond", "proper_cond"(default). For details, see vignette.

root_state_weight

the method to weigh the states:"maddison_weights", "proper_weights"(default) or "equal_weights". It can also be specified the root state:the vector c(1,0,0) indicates state 1 was the root state.

sampling_fraction

vector that states the sampling proportion per trait state. It must have as many elements as trait states.

setting_calculation

argument used internally to speed up calculation. It should be left blank (default : setting_calculation = NULL)

loglik_penalty

the size of the penalty for all parameters; default is 0 (no penalty)

is_complete_tree

whether or not a tree with all its extinct species is provided

atol

absolute tolerance of integration

rtol

relative tolerance of integration

method

integration method used, available are: "odeint::runge_kutta_cash_karp54", "odeint::runge_kutta_fehlberg78", "odeint::runge_kutta_dopri5", "odeint::bulirsch_stoer" and "odeint::runge_kutta4". Default method is:"odeint::bulirsch_stoer".

num_steps

number of substeps to show intermediate likelihoods along a branch, if left to NULL, the intermediate likelihoods at every integration evaluation are stored, which is more exact, but can lead to huge datasets / memory usage.

verbose

provides intermediate output if TRUE

Value

The loglikelihood of the data given the parameters

Examples

#' set.seed(5)
focal_tree <- ape::rphylo(n = 4, birth = 1, death = 0)
traits <- c(0, 1, 1, 0)
params <- secsse::id_paramPos(c(0, 1), 2)
params[[1]][] <- c(0.2, 0.2, 0.1, 0.1)
params[[2]][] <- 0.0
params[[3]][, ] <- 0.1
diag(params[[3]]) <- NA
#  Thus, we have for both, rates
# 0A, 1A, 0B and 1B. If we are interested in the posterior probability of
# trait 0 we have to provide a helper function that sums the probabilities of
# 0A and 0B, e.g.:
helper_function <- function(x) {
  return(sum(x[c(5, 7)]) / sum(x)) # normalized by total sum, just in case.
}
ll <- secsse::secsse_loglik(parameter = params,
                            phy = focal_tree,
                            traits = traits,
                            num_concealed_states = 2,
                            sampling_fraction = c(1, 1),
                            see_ancestral_states = TRUE)

secsse_loglik_eval(parameter = params,
                   phy = focal_tree,
                   traits = traits,
                   ancestral_states = ll$states,
                   num_concealed_states = 2,
                   sampling_fraction = c(1, 1),
                   num_steps = 10)
#>       [,1] [,2]        [,3] [,4] [,5] [,6] [,7]        [,8]       [,9]
#>  [1,]    7    2 0.000000000    0    0    0    0 0.000000000 1.00000000
#>  [2,]    7    2 0.061749385    0    0    0    0 0.012067596 0.97569590
#>  [3,]    7    2 0.123498770    0    0    0    0 0.023585127 0.95216464
#>  [4,]    7    2 0.185248155    0    0    0    0 0.034573732 0.92937924
#>  [5,]    7    2 0.246997541    0    0    0    0 0.045053766 0.90731368
#>  [6,]    7    2 0.308746926    0    0    0    0 0.055044830 0.88594290
#>  [7,]    7    2 0.370496311    0    0    0    0 0.064565797 0.86524270
#>  [8,]    7    2 0.432245696    0    0    0    0 0.073634839 0.84518976
#>  [9,]    7    2 0.493995081    0    0    0    0 0.082269453 0.82576162
#> [10,]    7    2 0.555744466    0    0    0    0 0.090486486 0.80693659
#> [11,]    7    2 0.617493852    0    0    0    0 0.098302161 0.78869376
#> [12,]    7    3 0.000000000    0    0    0    0 0.000000000 1.00000000
#> [13,]    7    3 0.061749385    0    0    0    0 0.012067596 0.97569590
#> [14,]    7    3 0.123498770    0    0    0    0 0.023585127 0.95216464
#> [15,]    7    3 0.185248155    0    0    0    0 0.034573732 0.92937924
#> [16,]    7    3 0.246997541    0    0    0    0 0.045053766 0.90731368
#> [17,]    7    3 0.308746926    0    0    0    0 0.055044830 0.88594290
#> [18,]    7    3 0.370496311    0    0    0    0 0.064565797 0.86524270
#> [19,]    7    3 0.432245696    0    0    0    0 0.073634839 0.84518976
#> [20,]    7    3 0.493995081    0    0    0    0 0.082269453 0.82576162
#> [21,]    7    3 0.555744466    0    0    0    0 0.090486486 0.80693659
#> [22,]    7    3 0.617493852    0    0    0    0 0.098302161 0.78869376
#> [23,]    6    7 0.000000000    0    0    0    0 0.009800922 0.63089575
#> [24,]    6    7 0.004461637    0    0    0    0 0.010220124 0.62965451
#> [25,]    6    7 0.008923273    0    0    0    0 0.010637879 0.62841627
#> [26,]    6    7 0.013384910    0    0    0    0 0.011054193 0.62718101
#> [27,]    6    7 0.017846547    0    0    0    0 0.011469068 0.62594873
#> [28,]    6    7 0.022308184    0    0    0    0 0.011882508 0.62471942
#> [29,]    6    7 0.026769820    0    0    0    0 0.012294519 0.62349307
#> [30,]    6    7 0.031231457    0    0    0    0 0.012705105 0.62226968
#> [31,]    6    7 0.035693094    0    0    0    0 0.013114268 0.62104923
#> [32,]    6    7 0.040154730    0    0    0    0 0.013522014 0.61983172
#> [33,]    6    7 0.044616367    0    0    0    0 0.013928347 0.61861715
#> [34,]    6    1 0.000000000    0    0    0    0 1.000000000 0.00000000
#> [35,]    6    1 0.066211022    0    0    0    0 0.973970104 0.01291796
#> [36,]    6    1 0.132422044    0    0    0    0 0.948826493 0.02520527
#> [37,]    6    1 0.198633066    0    0    0    0 0.924536041 0.03688789
#> [38,]    6    1 0.264844087    0    0    0    0 0.901066891 0.04799073
#> [39,]    6    1 0.331055109    0    0    0    0 0.878388409 0.05853774
#> [40,]    6    1 0.397266131    0    0    0    0 0.856471134 0.06855189
#> [41,]    6    1 0.463477153    0    0    0    0 0.835286733 0.07805526
#> [42,]    6    1 0.529688175    0    0    0    0 0.814807961 0.08706903
#> [43,]    6    1 0.595899197    0    0    0    0 0.795008614 0.09561355
#> [44,]    6    1 0.662110219    0    0    0    0 0.775863491 0.10370837
#> [45,]    5    6 0.000000000    0    0    0    0 0.110673162 0.65704211
#> [46,]    5    6 0.014955320    0    0    0    0 0.111171387 0.65265959
#> [47,]    5    6 0.029910641    0    0    0    0 0.111661239 0.64831230
#> [48,]    5    6 0.044865961    0    0    0    0 0.112142808 0.64399993
#> [49,]    5    6 0.059821282    0    0    0    0 0.112616183 0.63972220
#> [50,]    5    6 0.074776602    0    0    0    0 0.113081451 0.63547880
#> [51,]    5    6 0.089731923    0    0    0    0 0.113538698 0.63126944
#> [52,]    5    6 0.104687243    0    0    0    0 0.113988011 0.62709384
#> [53,]    5    6 0.119642564    0    0    0    0 0.114429474 0.62295170
#> [54,]    5    6 0.134597884    0    0    0    0 0.114863171 0.61884274
#> [55,]    5    6 0.149553205    0    0    0    0 0.115289185 0.61476668
#> [56,]    5    4 0.000000000    0    0    0    0 1.000000000 0.00000000
#> [57,]    5    4 0.081166342    0    0    0    0 0.968214545 0.01574753
#> [58,]    5    4 0.162332685    0    0    0    0 0.937749637 0.03055623
#> [59,]    5    4 0.243499027    0    0    0    0 0.908545047 0.04447325
#> [60,]    5    4 0.324665369    0    0    0    0 0.880543365 0.05754348
#> [61,]    5    4 0.405831712    0    0    0    0 0.853689868 0.06980963
#> [62,]    5    4 0.486998054    0    0    0    0 0.827932391 0.08131232
#> [63,]    5    4 0.568164396    0    0    0    0 0.803221208 0.09209022
#> [64,]    5    4 0.649330739    0    0    0    0 0.779508917 0.10218011
#> [65,]    5    4 0.730497081    0    0    0    0 0.756750331 0.11161699
#> [66,]    5    4 0.811663423    0    0    0    0 0.734902371 0.12043414
#>             [,10]      [,11]
#>  [1,] 0.000000000 1.00000000
#>  [2,] 0.012104758 0.98170182
#>  [3,] 0.023730007 0.96384847
#>  [4,] 0.034891462 0.94642756
#>  [5,] 0.045604344 0.92942707
#>  [6,] 0.055883398 0.91283532
#>  [7,] 0.065742903 0.89664097
#>  [8,] 0.075196694 0.88083302
#>  [9,] 0.084258169 0.86540078
#> [10,] 0.092940304 0.85033389
#> [11,] 0.101255670 0.83562227
#> [12,] 0.000000000 1.00000000
#> [13,] 0.012104758 0.98170182
#> [14,] 0.023730007 0.96384847
#> [15,] 0.034891462 0.94642756
#> [16,] 0.045604344 0.92942707
#> [17,] 0.055883398 0.91283532
#> [18,] 0.065742903 0.89664097
#> [19,] 0.075196694 0.88083302
#> [20,] 0.084258169 0.86540078
#> [21,] 0.092940304 0.85033389
#> [22,] 0.101255670 0.83562227
#> [23,] 0.005199355 0.35410398
#> [24,] 0.005633273 0.35376042
#> [25,] 0.006065897 0.35341730
#> [26,] 0.006497232 0.35307462
#> [27,] 0.006927281 0.35273238
#> [28,] 0.007356046 0.35239058
#> [29,] 0.007783531 0.35204921
#> [30,] 0.008209739 0.35170828
#> [31,] 0.008634673 0.35136779
#> [32,] 0.009058336 0.35102773
#> [33,] 0.009480732 0.35068810
#> [34,] 1.000000000 0.00000000
#> [35,] 0.980397088 0.01296061
#> [36,] 0.961304571 0.02537123
#> [37,] 0.942707226 0.03725117
#> [38,] 0.924590305 0.04861909
#> [39,] 0.906939523 0.05949301
#> [40,] 0.889741037 0.06989037
#> [41,] 0.872981440 0.07982799
#> [42,] 0.856647741 0.08932216
#> [43,] 0.840727355 0.09838860
#> [44,] 0.825208089 0.10704250
#> [45,] 0.040062017 0.19222272
#> [46,] 0.041251549 0.19227869
#> [47,] 0.042428294 0.19233031
#> [48,] 0.043592364 0.19237765
#> [49,] 0.044743866 0.19242073
#> [50,] 0.045882908 0.19245962
#> [51,] 0.047009597 0.19249434
#> [52,] 0.048124040 0.19252496
#> [53,] 0.049226341 0.19255150
#> [54,] 0.050316606 0.19257403
#> [55,] 0.051394938 0.19259258
#> [56,] 1.000000000 0.00000000
#> [57,] 0.976040522 0.01581122
#> [58,] 0.952842848 0.03080254
#> [59,] 0.930379232 0.04500912
#> [60,] 0.908622988 0.05846467
#> [61,] 0.887548446 0.07120152
#> [62,] 0.867130917 0.08325068
#> [63,] 0.847346651 0.09464188
#> [64,] 0.828172805 0.10540363
#> [65,] 0.809587402 0.11556327
#> [66,] 0.791569305 0.12514700