Maximum likehood estimation for (SecSSE)

Maximum likehood estimation under Several examined and concealed States-dependent Speciation and Extinction (SecSSE)

Usage

secsse_ml(
  phy,
  traits,
  num_concealed_states,
  idparslist,
  idparsopt,
  initparsopt,
  idparsfix,
  parsfix,
  cond = "proper_cond",
  root_state_weight = "proper_weights",
  sampling_fraction,
  tol = c(1e-04, 1e-05, 1e-07),
  maxiter = 1000 * round((1.25)^length(idparsopt)),
  optimmethod = "simplex",
  num_cycles = 1,
  loglik_penalty = 0,
  is_complete_tree = FALSE,
  take_into_account_root_edge = FALSE,
  verbose = FALSE,
  num_threads = 1,
  atol = 1e-08,
  rtol = 1e-07,
  method = "odeint::bulirsch_stoer",
  use_normalization = TRUE
)

Arguments

phy

phylogenetic tree of class phylo, rooted and with branch lengths. Alternatively, multiple phylogenetic trees can be provided as the multiPhylo class.

traits

vector with trait states for each tip in the phylogeny. The order of the states must be the same as the tree tips. For help, see vignette("starting_secsse", package = "secsse"). When providing a multiPhylo set of multiple phylognies, traits should be a list where each entry in the list corresponds to the matching phylogeny on that position.

num_concealed_states

number of concealed states, generally equivalent to the number of examined states in the dataset.

idparslist

overview of parameters and their values.

idparsopt

a numeric vector with the ID of parameters to be estimated.

initparsopt

a numeric vector with the initial guess of the parameters to be estimated.

idparsfix

a numeric vector with the ID of the fixed parameters.

parsfix

a numeric vector with the value of the fixed parameters.

cond

condition on the existence of a node root: "maddison_cond", "proper_cond" (default). For details, see vignette.

root_state_weight

the method to weigh the states: "maddison_weights", "proper_weights" (default) or "equal_weights". It can also be specified for the root state: the vector c(1, 0, 0) indicates state 1 was the root state. When using a multiPhylo object, root_state_weight should be list where each entry in the list corresponds to the root_state_weight for each tree.

sampling_fraction

vector that states the sampling proportion per trait state. It must have as many elements as there are trait states. When using a multiPhylo object, sampling fraction should be list where each entry in the list corresponds to the sampling proportion for each tree.

tol

A numeric vector with the maximum tolerance of the optimization algorithm. Default is c(1e-04, 1e-05, 1e-05).

maxiter

max number of iterations. Default is 1000 * round((1.25) ^ length(idparsopt)).

optimmethod

A string with method used for optimization. Default is "simplex". Alternative is "subplex". Both are called from DDD::optimizer(), simplex is implemented natively in DDD, while subplex is ultimately called from subplex::subplex().

num_cycles

Number of cycles of the optimization. When set to Inf, the optimization will be repeated until the result is, within the tolerance, equal to the starting values, with a maximum of 10 cycles.

loglik_penalty

the size of the penalty for all parameters; default is 0 (no penalty).

is_complete_tree

logical specifying whether or not a tree with all its extinct species is provided. If set to TRUE, it also assumes that all all extinct lineages are present on the tree. Defaults to FALSE.

take_into_account_root_edge

if TRUE, the LL integration is continued along the root edge. This also affects conditioning (as now, conditioning no longer needs to assume a speciation event at the start of the tree)

verbose

sets verbose output; default is TRUE when optimmethod is "simplex". If optimmethod is set to "simplex", then even if set to FALSE, optimizer output will be shown.

num_threads

number of threads to be used. Default is one thread.

atol

A numeric specifying the absolute tolerance of integration.

rtol

A numeric specifying the relative tolerance of integration.

method

integration method used, available are: "odeint::runge_kutta_cash_karp54", "odeint::runge_kutta_fehlberg78", "odeint::runge_kutta_dopri5", "odeint::bulirsch_stoer" and "odeint::runge_kutta4". Default method is: "odeint::bulirsch_stoer".

use_normalization

normalize the density vector during integration, more accurate but slower (default = TRUE)

Value

Parameter estimated and maximum likelihood

Examples

# Example of how to set the arguments for a ML search.
library(secsse)
library(DDD)
set.seed(13)
# lambdas for 0A and 1A and 2A are the same but need to be estimated
# mus are fixed to
# the transition rates are constrained to be equal and fixed 0.01
phylotree <- ape::rcoal(31, tip.label = 1:31)
traits <-  sample(c(0,1,2), ape::Ntip(phylotree),replace=TRUE)#get some traits
num_concealed_states<-3
idparslist <- id_paramPos(traits, num_concealed_states)
idparslist[[1]][c(1,4,7)] <- 1
idparslist[[1]][c(2,5,8)] <- 2
idparslist[[1]][c(3,6,9)] <- 3
idparslist[[2]][]<-4
masterBlock <- matrix(5,ncol = 3,nrow = 3,byrow = TRUE)
diag(masterBlock) <- NA
diff.conceal <- FALSE
idparslist[[3]] <- q_doubletrans(traits,masterBlock,diff.conceal)
startingpoint <- DDD::bd_ML(brts = ape::branching.times(phylotree))
#> You are optimizing lambda0 mu0 
#> You are fixing lambda1 mu1 
#> Optimizing the likelihood - this may take a while. 
#> The loglikelihood for the initial parameter values is -67.5669039297515.
#> 
#> Maximum likelihood parameter estimates: lambda0: 15.604290, mu0: 15.874464, lambda1: 0.000000, mu1: 0.000000: 
#> Maximum loglikelihood: 25.282018
intGuessLamba <- startingpoint$lambda0
intGuessMu <- startingpoint$mu0
idparsopt <- c(1,2,3,5)
initparsopt <- c(rep(intGuessLamba,3),rep((intGuessLamba/5),1))
idparsfix <- c(0,4)
parsfix <- c(0,0)
tol <- c(1e-02, 1e-03, 1e-04)
maxiter <- 1000 * round((1.25)^length(idparsopt))
optimmethod <- 'simplex'
cond <- 'proper_cond'
root_state_weight <- 'proper_weights'
sampling_fraction <- c(1,1,1)
model<-secsse_ml(
phylotree,
traits,
num_concealed_states,
idparslist,
idparsopt,
initparsopt,
idparsfix,
parsfix,
cond,
root_state_weight,
sampling_fraction,
tol,
maxiter,
optimmethod,
num_cycles = 1,
verbose = FALSE)
#> Warning: Note: you set some transitions as impossible to happen.
#> 1 15.6042898684248 15.6042898684248 15.6042898684248 3.12085797368496 -55.0829332014493 initial 
#> 2 16.1826565594201 16.1826565594201 14.2366516329269 3.23679263471176 -53.5955770049681 expand 
#> 3 16.1826565594201 16.1826565594201 14.2366516329269 3.23679263471176 -53.5955770049681 reflect 
#> 4 15.6744474204993 15.0418248706591 14.0813764954283 3.43976602145215 -51.0733820765387 expand 
#> 5 15.6744474204993 15.0418248706591 14.0813764954283 3.43976602145215 -51.0733820765387 reflect 
#> 6 15.7897266624575 16.3431211350014 12.1251758810308 3.51957331290552 -48.3309396326738 expand 
#> 7 15.7897266624575 16.3431211350014 12.1251758810308 3.51957331290552 -48.3309396326738 reflect 
#> 8 15.4239724497177 14.7480876746975 11.1658804598539 4.18949505076817 -43.7404820698376 expand 
#> 9 15.4239724497177 14.7480876746975 11.1658804598539 4.18949505076817 -43.7404820698376 reflect 
#> 10 16.5037295388755 15.9290391482382 8.60698610560038 4.62942837051011 -40.0935229627584 expand 
#> 11 16.5037295388755 15.9290391482382 8.60698610560038 4.62942837051011 -40.0935229627584 reflect 
#> 12 15.8415238453205 13.3794084219105 7.17017313753921 6.95418165007756 -34.1209262551831 expand 
#> 13 15.8415238453205 13.3794084219105 7.17017313753921 6.95418165007756 -34.1209262551831 reflect 
#> 14 15.8415238453205 13.3794084219105 7.17017313753921 6.95418165007756 -34.1209262551831 reflect 
#> 15 17.5043421500954 13.3823278970245 4.87199464662075 18.4611852376756 -32.8526429864438 expand 
#> 16 15.9467686168646 12.5755678121138 4.78859464019957 27.3897761206681 -30.498362602307 reflect 
#> 17 15.9467686168646 12.5755678121138 4.78859464019957 27.3897761206681 -30.498362602307 reflect 
#> 18 15.9467686168646 12.5755678121138 4.78859464019957 27.3897761206681 -30.498362602307 reflect 
#> 19 15.9467686168646 12.5755678121138 4.78859464019957 27.3897761206681 -30.498362602307 contract inside 
#> 20 15.147576617686 10.5613404064817 4.69711763222928 94.2610341671819 -28.3896430421991 expand 
#> 21 15.147576617686 10.5613404064817 4.69711763222928 94.2610341671819 -28.3896430421991 contract inside 
#> 22 15.147576617686 10.5613404064817 4.69711763222928 94.2610341671819 -28.3896430421991 reflect 
#> 23 13.2944179344031 11.785197674832 6.05959674478364 12.5257381509452 -27.9360241354328 expand 
#> 24 12.427159941311 10.025417328563 5.11292486227123 201.453133415635 -25.6515040381369 expand 
#> 25 10.9905939475982 8.34578335972368 6.17006883864307 90.4771031319024 -23.1011245447053 expand 
#> 26 9.75994290755567 8.57564220737894 6.49398223369463 26.0606430096442 -21.6551080808365 expand 
#> 27 9.75994290755567 8.57564220737894 6.49398223369463 26.0606430096442 -21.6551080808365 reflect 
#> 28 9.75994290755567 8.57564220737894 6.49398223369463 26.0606430096442 -21.6551080808365 contract inside 
#> 29 9.75994290755567 8.57564220737894 6.49398223369463 26.0606430096442 -21.6551080808365 reflect 
#> 30 8.10099790792233 7.09013800751324 8.67472282764109 27.1921182249967 -20.769662199107 reflect 
#> 31 8.10099790792233 7.09013800751324 8.67472282764109 27.1921182249967 -20.769662199107 contract inside 
#> 32 8.10099790792233 7.09013800751324 8.67472282764109 27.1921182249967 -20.769662199107 reflect 
#> 33 8.10099790792233 7.09013800751324 8.67472282764109 27.1921182249967 -20.769662199107 reflect 
#> 34 8.10099790792233 7.09013800751324 8.67472282764109 27.1921182249967 -20.769662199107 reflect 
#> 35 7.6296917700292 6.57122597333637 7.92660128042042 32.8090393064141 -19.4764162710395 expand 
#> 36 7.6296917700292 6.57122597333637 7.92660128042042 32.8090393064141 -19.4764162710395 contract inside 
#> 37 7.6296917700292 6.57122597333637 7.92660128042042 32.8090393064141 -19.4764162710395 reflect 
#> 38 7.6296917700292 6.57122597333637 7.92660128042042 32.8090393064141 -19.4764162710395 reflect 
#> 39 7.33360372296142 5.97316058830953 7.68427636024989 118.816682619588 -19.4198777455986 reflect 
#> 40 7.33360372296142 5.97316058830953 7.68427636024989 118.816682619588 -19.4198777455986 reflect 
#> 41 6.82280472299841 5.92679948615982 7.53991257390515 158.727081937072 -19.0941235235379 reflect 
#> 42 6.56993279361279 5.43595841112662 7.97018235660804 38.4164204484102 -18.1404629506757 expand 
#> 43 6.56993279361279 5.43595841112662 7.97018235660804 38.4164204484102 -18.1404629506757 reflect 
#> 44 6.56993279361279 5.43595841112662 7.97018235660804 38.4164204484102 -18.1404629506757 contract inside 
#> 45 6.56993279361279 5.43595841112662 7.97018235660804 38.4164204484102 -18.1404629506757 reflect 
#> 46 6.56993279361279 5.43595841112662 7.97018235660804 38.4164204484102 -18.1404629506757 reflect 
#> 47 6.56993279361279 5.43595841112662 7.97018235660804 38.4164204484102 -18.1404629506757 reflect 
#> 48 5.28548785597528 4.55759418497378 9.04362818071381 21.1941663790621 -17.6156304997858 expand 
#> 49 7.02879026644281 5.68729611088657 5.91795875462147 16.7956725521271 -17.4832384474098 expand 
#> 50 6.0027044935016 5.07393969003788 5.8849948419789 32.6143423991358 -16.7040165369596 expand 
#> 51 6.0027044935016 5.07393969003788 5.8849948419789 32.6143423991358 -16.7040165369596 reflect 
#> 52 6.0027044935016 5.07393969003788 5.8849948419789 32.6143423991358 -16.7040165369596 reflect 
#> 53 6.59731881865314 5.16880316840877 4.59242490608013 16.5576292279009 -16.6550834529531 reflect 
#> 54 4.89250652444853 4.03050834801968 5.62433227241967 20.6451732738895 -16.2362567251028 reflect 
#> 55 4.89250652444853 4.03050834801968 5.62433227241967 20.6451732738895 -16.2362567251028 reflect 
#> 56 4.89250652444853 4.03050834801968 5.62433227241967 20.6451732738895 -16.2362567251028 reflect 
#> 57 4.89250652444853 4.03050834801968 5.62433227241967 20.6451732738895 -16.2362567251028 reflect 
#> 58 4.89250652444853 4.03050834801968 5.62433227241967 20.6451732738895 -16.2362567251028 reflect 
#> 59 4.89250652444853 4.03050834801968 5.62433227241967 20.6451732738895 -16.2362567251028 contract inside 
#> 60 4.89250652444853 4.03050834801968 5.62433227241967 20.6451732738895 -16.2362567251028 contract inside 
#> 61 4.89250652444853 4.03050834801968 5.62433227241967 20.6451732738895 -16.2362567251028 reflect 
#> 62 5.36292098261263 4.29755626460154 4.9432116196164 20.9554132411357 -16.2305966297634 contract outside 
#> 63 5.43119190496216 4.42130450226141 4.63166814110711 24.5958372312959 -16.222100655416 contract inside 
#> 64 4.7322264841979 3.85798700599787 6.07104455700087 29.2728976901882 -16.2133948951095 reflect 
#> 65 4.91809644345243 3.99201132494822 5.06667446879964 26.1345881287268 -16.196587283697 contract inside 
#> 66 4.91809644345243 3.99201132494822 5.06667446879964 26.1345881287268 -16.196587283697 contract inside 
#> 67 4.91809644345243 3.99201132494822 5.06667446879964 26.1345881287268 -16.196587283697 reflect 
#> 68 5.11626629597035 4.17542026220493 5.03471485009209 25.7817285958635 -16.1950921061624 contract inside 
#> 69 5.06776822078151 4.09944794960308 5.25844675572359 23.3631227319687 -16.1940413728444 contract outside 
#> 70 4.87393343865206 3.96942443425438 5.59519716752651 26.6069763274549 -16.1883991958061 contract inside 
#> 71 4.87393343865206 3.96942443425438 5.59519716752651 26.6069763274549 -16.1883991958061 contract inside 
#> 72 4.87393343865206 3.96942443425438 5.59519716752651 26.6069763274549 -16.1883991958061 contract inside 
#> 73 4.87393343865206 3.96942443425438 5.59519716752651 26.6069763274549 -16.1883991958061 reflect 
#> 74 4.77262513746732 3.89796625572618 5.60040120501363 26.6663710202517 -16.1854159733015 reflect 
#> 75 4.77262513746732 3.89796625572618 5.60040120501363 26.6663710202517 -16.1854159733015 contract inside 
#> 76 4.77262513746732 3.89796625572618 5.60040120501363 26.6663710202517 -16.1854159733015 contract inside 
#> 77 4.91209832956955 3.99632097953225 5.3466965313345 25.5730292012906 -16.1853515907983 contract inside 
#> 78 4.8684042341106 3.96366018361655 5.39995406717573 24.8674024723561 -16.1846887545063 contract outside 
#> 79 4.86939969565057 3.96095533029654 5.48048289360783 25.0778517814404 -16.1846673311355 contract inside 
#> 80 4.86939969565057 3.96095533029654 5.48048289360783 25.0778517814404 -16.1846673311355 reflect 
#> 81 4.81480199115496 3.92563417491044 5.51825896546419 26.0353970092068 -16.1843444806998 contract inside 
#> 82 4.81480199115496 3.92563417491044 5.51825896546419 26.0353970092068 -16.1843444806998 reflect 
#> 83 4.80442051858571 3.91305955878167 5.51631017722255 25.8035376268681 -16.1841903614628 contract inside 
#> 84 4.83960328595387 3.94124851189487 5.4661965593553 25.2304664626269 -16.184117396184 contract inside 
#> 85 4.83960328595387 3.94124851189487 5.4661965593553 25.2304664626269 -16.184117396184 contract inside 
#> 86 4.79009685844859 3.90314059407521 5.56113479702454 25.560280848137 -16.1840528770147 contract inside 
#> 87 4.79009685844859 3.90314059407521 5.56113479702454 25.560280848137 -16.1840528770147 contract inside 
#> 88 4.79009685844859 3.90314059407521 5.56113479702454 25.560280848137 -16.1840528770147 contract inside 
#> 89 4.79282808766046 3.90698915212831 5.52269364406131 25.7357051253013 -16.1840371812808 reflect 
#> 90 4.8211760660312 3.92740183197262 5.49704836790131 25.4497979455624 -16.183994155229 contract inside 
#> 91 4.8211760660312 3.92740183197262 5.49704836790131 25.4497979455624 -16.183994155229 reflect 
#> 92 4.78565141183831 3.90096240033502 5.54238582920928 25.4517694533707 -16.1839802169143 reflect 
#> 93 4.78565141183831 3.90096240033502 5.54238582920928 25.4517694533707 -16.1839802169143 reflect 
#> 94 4.78565141183831 3.90096240033502 5.54238582920928 25.4517694533707 -16.1839802169143 contract inside 
#> 95 4.81265691520433 3.92321817841846 5.49000194047387 25.5498069297059 -16.1839476623922 reflect 
#> 96 4.81265691520433 3.92321817841846 5.49000194047387 25.5498069297059 -16.1839476623922 reflect 
#> 97 4.81265691520433 3.92321817841846 5.49000194047387 25.5498069297059 -16.1839476623922 reflect 
#> 98 4.78987270355293 3.90682029657606 5.55085462395537 25.4411289935749 -16.1838937513851 expand 
#> 99 4.80571214245693 3.91934757404329 5.52042309257121 25.6285622976534 -16.1838766460877 reflect 
#> 100 4.79054155532026 3.90891204000868 5.51644500048978 25.5728661089364 -16.1838495116241 reflect 
#> 101 4.82171211098813 3.9326946175437 5.49161860359742 25.474434602463 -16.1838433916303 reflect 
#> 102 4.79124326218238 3.91064241962564 5.54980733801109 25.5082969712271 -16.1837896715737 reflect 
#> 103 4.79124326218238 3.91064241962564 5.54980733801109 25.5082969712271 -16.1837896715737 reflect 
#> 104 4.80214712396877 3.92216783726696 5.49369751117968 25.3990246019667 -16.1837055700874 expand 
#> 105 4.80214712396877 3.92216783726696 5.49369751117968 25.3990246019667 -16.1837055700874 reflect 
#> 106 4.80214712396877 3.92216783726696 5.49369751117968 25.3990246019667 -16.1837055700874 reflect 
#> 107 4.80214712396877 3.92216783726696 5.49369751117968 25.3990246019667 -16.1837055700874 reflect 
#> 108 4.80810662624832 3.93052935231345 5.51647989738955 25.4505968900043 -16.1835424476385 expand 
#> 109 4.80810662624832 3.93052935231345 5.51647989738955 25.4505968900043 -16.1835424476385 reflect 
#> 110 4.80810662624832 3.93052935231345 5.51647989738955 25.4505968900043 -16.1835424476385 reflect 
#> 111 4.80810662624832 3.93052935231345 5.51647989738955 25.4505968900043 -16.1835424476385 reflect 
#> 112 4.80810662624832 3.93052935231345 5.51647989738955 25.4505968900043 -16.1835424476385 reflect 
#> 113 4.80291201467036 3.9380830817591 5.51193974240646 25.2851662684525 -16.1832224797021 expand 
#> 114 4.80291201467036 3.9380830817591 5.51193974240646 25.2851662684525 -16.1832224797021 reflect 
#> 115 4.80291201467036 3.9380830817591 5.51193974240646 25.2851662684525 -16.1832224797021 reflect 
#> 116 4.80291201467036 3.9380830817591 5.51193974240646 25.2851662684525 -16.1832224797021 reflect 
#> 117 4.8118472284776 3.96266974335116 5.50137636537717 25.4676873442277 -16.1826802659679 expand 
#> 118 4.8118472284776 3.96266974335116 5.50137636537717 25.4676873442277 -16.1826802659679 reflect 
#> 119 4.8118472284776 3.96266974335116 5.50137636537717 25.4676873442277 -16.1826802659679 reflect 
#> 120 4.78938682734194 3.96478770482327 5.53004077559544 25.0177592529077 -16.1825439215219 expand 
#> 121 4.76600606211963 3.96460169211207 5.54727307758642 25.3662364888985 -16.1815042260551 expand 
#> 122 4.76600606211963 3.96460169211207 5.54727307758642 25.3662364888985 -16.1815042260551 reflect 
#> 123 4.76600606211963 3.96460169211207 5.54727307758642 25.3662364888985 -16.1815042260551 reflect 
#> 124 4.76718885570213 4.00088590677792 5.53777950201381 24.9218598327399 -16.1814463976082 reflect 
#> 125 4.77285688369093 4.04347163843464 5.50434760501419 25.4724154269505 -16.1796421161283 expand 
#> 126 4.77285688369093 4.04347163843464 5.50434760501419 25.4724154269505 -16.1796421161283 reflect 
#> 127 4.69577523668416 4.00670982446518 5.57594170601207 25.4225816257582 -16.1785177629163 expand 
#> 128 4.69577523668416 4.00670982446518 5.57594170601207 25.4225816257582 -16.1785177629163 reflect 
#> 129 4.67229066084091 4.09910002252224 5.58959264101893 26.1477450930662 -16.1752543527274 expand 
#> 130 4.67229066084091 4.09910002252224 5.58959264101893 26.1477450930662 -16.1752543527274 reflect 
#> 131 4.54460357728952 4.11540462842328 5.67870954190521 25.989050723082 -16.171471835289 expand 
#> 132 4.54460357728952 4.11540462842328 5.67870954190521 25.989050723082 -16.171471835289 reflect 
#> 133 4.4703205683801 4.29593108965953 5.66625400454508 27.7226727400669 -16.1647135347209 expand 
#> 134 4.4703205683801 4.29593108965953 5.66625400454508 27.7226727400669 -16.1647135347209 reflect 
#> 135 4.20893158867913 4.33178043019349 5.87960755733134 28.5427284205482 -16.1547155270114 expand 
#> 136 4.20893158867913 4.33178043019349 5.87960755733134 28.5427284205482 -16.1547155270114 reflect 
#> 137 4.16045295141311 4.5076682655382 5.90775728251757 30.2206907297828 -16.1542825154999 reflect 
#> 138 3.9553584975409 4.77438936188363 5.96819217138372 31.5742192773714 -16.14680375018 expand 
#> 139 3.85553006629305 4.7115462485019 6.16062038760346 31.6834915950822 -16.146531987297 reflect 
#> 140 3.83544244416796 4.7745763524055 6.09475914820699 33.0813847654062 -16.1459559184847 reflect 
#> 141 3.83544244416796 4.7745763524055 6.09475914820699 33.0813847654062 -16.1459559184847 contract inside 
#> 142 3.71667613453249 4.88308714306479 6.18469424268695 32.993316036119 -16.1421906839418 reflect 
#> 143 3.71667613453249 4.88308714306479 6.18469424268695 32.993316036119 -16.1421906839418 contract inside 
#> 144 3.71667613453249 4.88308714306479 6.18469424268695 32.993316036119 -16.1421906839418 reflect 
#> 145 3.71213034006565 4.84590512597595 6.10846801575376 34.2259455550035 -16.1399749217061 expand 
#> 146 3.66056964767875 4.77043027674806 6.25265170673819 32.1311737665243 -16.1354383280813 expand 
#> 147 3.66056964767875 4.77043027674806 6.25265170673819 32.1311737665243 -16.1354383280813 reflect 
#> 148 3.47270121234705 5.17211939751399 6.15649077276535 35.2026702574825 -16.131780517034 expand 
#> 149 3.32471108739592 5.05242678942531 6.30731131725968 36.6876513487075 -16.1265280756181 expand 
#> 150 3.66051733513931 4.90655552451937 5.86400159206824 33.1507990397049 -16.1242629101283 expand 
#> 151 3.19251542538658 5.23977734804887 6.2067665320728 34.1623646040023 -16.1059745407254 expand 
#> 152 3.19251542538658 5.23977734804887 6.2067665320728 34.1623646040023 -16.1059745407254 reflect 
#> 153 3.0713601433538 5.12057297047307 5.96960259295406 35.6787968749244 -16.0980442492294 expand 
#> 154 3.0713601433538 5.12057297047307 5.96960259295406 35.6787968749244 -16.0980442492294 reflect 
#> 155 2.75799232321647 5.69008637928015 6.09401139923378 37.6610411818517 -16.0872622671625 reflect 
#> 156 2.75799232321647 5.69008637928015 6.09401139923378 37.6610411818517 -16.0872622671625 reflect 
#> 157 2.75799232321647 5.69008637928015 6.09401139923378 37.6610411818517 -16.0872622671625 reflect 
#> 158 2.36907583566798 5.4651883938382 6.36313502406932 39.4699318015767 -16.0741133848396 expand 
#> 159 2.12880486230196 6.19499640555913 6.14482346588648 37.5507635135595 -16.0643695120008 expand 
#> 160 2.12880486230196 6.19499640555913 6.14482346588648 37.5507635135595 -16.0643695120008 reflect 
#> 161 2.12880486230196 6.19499640555913 6.14482346588648 37.5507635135595 -16.0643695120008 contract inside 
#> 162 1.73441219380976 5.83311930160187 6.81652851389571 37.2149726223669 -16.0524093938854 expand 
#> 163 1.73441219380976 5.83311930160187 6.81652851389571 37.2149726223669 -16.0524093938854 reflect 
#> 164 1.73441219380976 5.83311930160187 6.81652851389571 37.2149726223669 -16.0524093938854 reflect 
#> 165 1.48937178227295 6.44108841195335 6.46930090538646 37.5603571884601 -16.0495564041272 reflect 
#> 166 1.48937178227295 6.44108841195335 6.46930090538646 37.5603571884601 -16.0495564041272 contract inside 
#> 167 1.48937178227295 6.44108841195335 6.46930090538646 37.5603571884601 -16.0495564041272 reflect 
#> 168 1.42553122356925 5.90541331486893 6.94002965964265 39.3947596430624 -16.0485845703539 reflect 
#> 169 1.710579993563 5.94327931131807 6.47838096736999 37.0126867781341 -16.0483604860391 contract inside 
#> 170 1.32837811499052 6.27361226139581 6.48883616784047 37.3221604482178 -16.0450057622495 reflect 
#> 171 1.32837811499052 6.27361226139581 6.48883616784047 37.3221604482178 -16.0450057622495 contract outside 
#> 172 1.32837811499052 6.27361226139581 6.48883616784047 37.3221604482178 -16.0450057622495 contract inside 
#> 173 1.46097184921084 6.03881681478332 6.7168164156165 38.5956898256667 -16.044388231279 contract inside 
#> 174 1.46097184921084 6.03881681478332 6.7168164156165 38.5956898256667 -16.044388231279 reflect 
#> 175 1.25503804352285 6.27695216564194 6.6750208234698 37.4908007340115 -16.0428199477253 reflect 
#> 176 1.02922663641634 6.27359589317912 6.83418182851971 38.9060891942224 -16.0414705481151 expand 
#> 177 1.02922663641634 6.27359589317912 6.83418182851971 38.9060891942224 -16.0414705481151 contract inside 
#> 178 1.02922663641634 6.27359589317912 6.83418182851971 38.9060891942224 -16.0414705481151 contract inside 
#> 179 1.02922663641634 6.27359589317912 6.83418182851971 38.9060891942224 -16.0414705481151 contract outside 
#> 180 1.02922663641634 6.27359589317912 6.83418182851971 38.9060891942224 -16.0414705481151 reflect 
#> 181 1.02922663641634 6.27359589317912 6.83418182851971 38.9060891942224 -16.0414705481151 reflect 
#> 182 1.02922663641634 6.27359589317912 6.83418182851971 38.9060891942224 -16.0414705481151 reflect 
#> 183 1.02922663641634 6.27359589317912 6.83418182851971 38.9060891942224 -16.0414705481151 contract inside 
#> 184 1.02612880593112 6.39589959996284 6.74969956924587 38.6945130060109 -16.0413973871668 contract inside 
#> 185 1.02612880593112 6.39589959996284 6.74969956924587 38.6945130060109 -16.0413973871668 contract inside 
#> 186 0.95749669486443 6.3595568186875 6.85106104359359 38.8134858114725 -16.0413932776777 contract outside 
#> 187 0.978051506384182 6.38061004246188 6.81858855978037 39.1284803987734 -16.0413642666585 contract inside 
#> 188 1.00076740298927 6.32695710131365 6.8230232925267 38.9311714310517 -16.0413266283329 contract inside 
#> 189 1.00076740298927 6.32695710131365 6.8230232925267 38.9311714310517 -16.0413266283329 reflect 
#> 190 1.01106667208897 6.37412555640932 6.78523488605711 38.7799946647196 -16.0413004516896 contract inside 
#> 191 1.01106667208897 6.37412555640932 6.78523488605711 38.7799946647196 -16.0413004516896 reflect 
#> 192 1.05953618730698 6.33351932144261 6.77531405707997 38.6395913826234 -16.0412941838268 contract outside 
#> 193 1.05953618730698 6.33351932144261 6.77531405707997 38.6395913826234 -16.0412941838268 reflect 
#> 194 1.01912818531399 6.33979766702619 6.79873982113646 38.8846365003985 -16.0412769507812 contract inside 
#> 195 1.01912818531399 6.33979766702619 6.79873982113646 38.8846365003985 -16.0412769507812 contract inside 
#> 196 1.01912818531399 6.33979766702619 6.79873982113646 38.8846365003985 -16.0412769507812 contract inside 
#> 197 1.05200010801089 6.32849984932135 6.77926211610369 38.8450794194844 -16.0412760691044 contract outside 
#> 198 1.0478630547679 6.3379502687977 6.77877557058843 38.7569520847271 -16.0412733714412 contract inside 
#> 199 1.0478630547679 6.3379502687977 6.77877557058843 38.7569520847271 -16.0412733714412 reflect 
#> 200 1.04524878068349 6.34233195775871 6.77855212630361 38.8315149394739 -16.0412709335214 contract inside 
#> 201 1.03451605938053 6.33704159662514 6.78954601244841 38.8405640495117 -16.0412691937031 contract inside 
#> 202 1.03451605938053 6.33704159662514 6.78954601244841 38.8405640495117 -16.0412691937031 reflect 
#> 203 1.03451605938053 6.33704159662514 6.78954601244841 38.8405640495117 -16.0412691937031 reflect 
#> 204 1.03451605938053 6.33704159662514 6.78954601244841 38.8405640495117 -16.0412691937031 contract inside 
#> 205 1.03451605938053 6.33704159662514 6.78954601244841 38.8405640495117 -16.0412691937031 reflect 
#> 206 1.03451605938053 6.33704159662514 6.78954601244841 38.8405640495117 -16.0412691937031 reflect 
#> 207 1.03329021870569 6.34440687504978 6.78683362342384 38.854492023612 -16.0412683770498 contract inside 
#> 208 1.03329021870569 6.34440687504978 6.78683362342384 38.854492023612 -16.0412683770498 reflect 
#> 209 1.03740919130229 6.34072436291063 6.78734196027895 38.809736919109 -16.0412679412403 contract inside 
#> 210 1.03740919130229 6.34072436291063 6.78734196027895 38.809736919109 -16.0412679412403 contract inside 
#> 211 1.03740919130229 6.34072436291063 6.78734196027895 38.809736919109 -16.0412679412403 contract inside 
#> 212 1.03740919130229 6.34072436291063 6.78734196027895 38.809736919109 -16.0412679412403 reflect 
#> 213 1.03517112301048 6.34202200845204 6.78794620616026 38.839409665135 -16.0412677257959 contract inside 
#> 214 1.03517112301048 6.34202200845204 6.78794620616026 38.839409665135 -16.0412677257959 contract inside 
#> 215 1.03555848514655 6.33909090835021 6.79013328464027 38.8284800224887 -16.0412676957092 contract inside 
#> 216 1.03555848514655 6.33909090835021 6.79013328464027 38.8284800224887 -16.0412676957092 contract inside 
#> 217 1.03690219856286 6.34050203801758 6.78807513574642 38.820366395072 -16.0412676913132 contract inside 
#> 218 1.03667941466109 6.34064583340097 6.78860861959907 38.8277713268737 -16.0412676320537 contract inside 
#> 219 1.03667941466109 6.34064583340097 6.78860861959907 38.8277713268737 -16.0412676320537 contract inside 
#> 220 1.03667941466109 6.34064583340097 6.78860861959907 38.8277713268737 -16.0412676320537 reflect 
#> 221 1.035807196666 6.34001823689094 6.78932458296968 38.8274699020965 -16.0412676193386 contract inside 
#> 222 1.035807196666 6.34001823689094 6.78932458296968 38.8274699020965 -16.0412676193386 reflect 
#> 223 1.035807196666 6.34001823689094 6.78932458296968 38.8274699020965 -16.0412676193386 contract outside 
#> 224 1.03577606664144 6.34008398278648 6.78984400641904 38.8297837958867 -16.0412676134535 reflect 
#> 225 1.03540510783189 6.3406852709877 6.7893857686384 38.832973028128 -16.0412676119557 contract inside 
#> 226 1.03540510783189 6.3406852709877 6.7893857686384 38.832973028128 -16.0412676119557 reflect 
#> 227 1.03615161340261 6.34049295153755 6.78907353507191 38.8282480087805 -16.0412676077634 contract inside 
#> 228 1.03564648618698 6.34044647790106 6.78950332423485 38.8271456401929 -16.0412676064258 contract inside 
#> 229 1.03564648618698 6.34044647790106 6.78950332423485 38.8271456401929 -16.0412676064258 contract inside 
#> 230 1.03576041487999 6.34027290910557 6.78959083626006 38.8295006274034 -16.0412676054097 contract inside 
#> Optimization has terminated successfully. 
# model$ML
# [1] -16.04127