Maximum likehood estimation for (SecSSE) with parameter as complex functions.

Maximum likehood estimation under Several examined and concealed States-dependent Speciation and Extinction (SecSSE) where some paramaters are functions of other parameters and/or factors.

Usage

secsse_ml_func_def_pars(
  phy,
  traits,
  num_concealed_states,
  idparslist,
  idparsopt,
  initparsopt,
  idfactorsopt,
  initfactors,
  idparsfix,
  parsfix,
  idparsfuncdefpar,
  functions_defining_params,
  cond = "proper_cond",
  root_state_weight = "proper_weights",
  sampling_fraction,
  tol = c(1e-04, 1e-05, 1e-07),
  maxiter = 1000 * round((1.25)^length(idparsopt)),
  optimmethod = "simplex",
  num_cycles = 1,
  loglik_penalty = 0,
  is_complete_tree = FALSE,
  num_threads = 1,
  atol = 1e-08,
  rtol = 1e-06,
  method = "odeint::bulirsch_stoer"
)

Arguments

phy

phylogenetic tree of class phylo, ultrametric, rooted and with branch lengths.

traits

a vector with trait states for each tip in the phylogeny.

num_concealed_states

number of concealed states, generally equivalent to the number of examined states in the dataset.

idparslist

overview of parameters and their values.

idparsopt

id of parameters to be estimated.

initparsopt

initial guess of the parameters to be estimated.

idfactorsopt

id of the factors that will be optimized. There are not fixed factors, so use a constant within 'functions_defining_params'.

initfactors

the initial guess for a factor (it should be set to NULL when no factors).

idparsfix

id of the fixed parameters (it should be set to NULL when there are no factors).

parsfix

value of the fixed parameters.

idparsfuncdefpar

id of the parameters which will be a function of optimized and/or fixed parameters. The order of id should match functions_defining_params

functions_defining_params

a list of functions. Each element will be a function which defines a parameter e.g. id_3 <- (id_1+id_2)/2. See example and vigenette

cond

condition on the existence of a node root: "maddison_cond","proper_cond"(default). For details, see vignette.

root_state_weight

the method to weigh the states: "maddison_weights","proper_weights"(default) or "equal_weights". It can also be specified the root state:the vector c(1, 0, 0) indicates state 1 was the root state.

sampling_fraction

vector that states the sampling proportion per trait state. It must have as many elements as there are trait states.

tol

maximum tolerance. Default is "c(1e-04, 1e-05, 1e-05)".

maxiter

max number of iterations. Default is "1000 *round((1.25)^length(idparsopt))".

optimmethod

method used for optimization. Default is "simplex".

num_cycles

number of cycles of the optimization (default is 1).

loglik_penalty

the size of the penalty for all parameters; default is 0 (no penalty)

is_complete_tree

whether or not a tree with all its extinct species is provided

num_threads

number of threads. Set to -1 to use all available threads. Default is one thread.

atol

absolute tolerance of integration

rtol

relative tolerance of integration

method

integration method used, available are: "odeint::runge_kutta_cash_karp54", "odeint::runge_kutta_fehlberg78", "odeint::runge_kutta_dopri5", "odeint::bulirsch_stoer" and "odeint::runge_kutta4". Default method is:"odeint::bulirsch_stoer".

Value

Parameter estimated and maximum likelihood

Parameter estimated and maximum likelihood

Examples

# Example of how to set the arguments for a ML search.
rm(list=ls(all=TRUE))
library(secsse)
library(DDD)
set.seed(16)
phylotree <- ape::rbdtree(0.07,0.001,Tmax=50)
startingpoint<-bd_ML(brts = ape::branching.times(phylotree))
#> You are optimizing lambda0 mu0 
#> You are fixing lambda1 mu1 
#> Optimizing the likelihood - this may take a while. 
#> The loglikelihood for the initial parameter values is -106.425581129189.
#> 
#> Maximum likelihood parameter estimates: lambda0: 0.056537, mu0: 0.000000, lambda1: 0.000000, mu1: 0.000000: 
#> Maximum loglikelihood: -104.592844
intGuessLamba <- startingpoint$lambda0
intGuessMu <- startingpoint$mu0
traits <- sample(c(0,1,2), ape::Ntip(phylotree),replace=TRUE) #get some traits
num_concealed_states<-3
idparslist<-id_paramPos(traits, num_concealed_states)
idparslist[[1]][c(1,4,7)] <- 1
idparslist[[1]][c(2,5,8)] <- 2
idparslist[[1]][c(3,6,9)] <- 3
idparslist[[2]][] <- 4
masterBlock <- matrix(c(5,6,5,6,5,6,5,6,5),ncol = 3,nrow = 3,byrow = TRUE)
diag(masterBlock) <- NA
diff.conceal <- FALSE
idparslist[[3]] <- q_doubletrans(traits,masterBlock,diff.conceal)
idparsfuncdefpar <- c(3,5,6)
idparsopt <- c(1,2)
idparsfix <- c(0,4)
initparsopt <- c(rep(intGuessLamba,2))
parsfix <- c(0,0)
idfactorsopt <- 1
initfactors <- 4
# functions_defining_params is a list of functions. Each function has no
# arguments and to refer
# to parameters ids should be indicated as "par_" i.e. par_3 refers to
# parameter 3. When a function is defined, be sure that all the parameters
# involved are either estimated, fixed or
# defined by previous functions (i.e, a function that defines parameter in
# 'functions_defining_params'). The user is responsible for this. In this
# exampl3, par_3 (i.e., parameter 3) is needed to calculate par_6. This is
# correct because par_3 is defined in
# the first function of 'functions_defining_params'. Notice that factor_1
# indicates a value that will be estimated to satisfy the equation. The same
# factor can be shared to define several parameters.
functions_defining_params <- list()
functions_defining_params[[1]] <- function(){
 par_3 <- par_1 + par_2
}
functions_defining_params[[2]] <- function(){
 par_5 <- par_1 * factor_1
}
functions_defining_params[[3]] <- function(){
 par_6 <- par_3 * factor_1
}

tol = c(1e-02, 1e-03, 1e-04)
maxiter = 1000 * round((1.25)^length(idparsopt))
optimmethod = "subplex"
cond<-"proper_cond"
root_state_weight <- "proper_weights"
sampling_fraction <- c(1,1,1)
model <- secsse_ml_func_def_pars(phylotree,
traits,
num_concealed_states,
idparslist,
idparsopt,
initparsopt,
idfactorsopt,
initfactors,
idparsfix,
parsfix,
idparsfuncdefpar,
functions_defining_params,
cond,
root_state_weight,
sampling_fraction,
tol,
maxiter,
optimmethod,
num_cycles = 1)
#> Warning: You set some transitions as impossible to happen
#> 0.0565371031208431 0.0565371031208431 4 -138.29892658976
#> Calculating the likelihood for the initial parameters.
#> The loglikelihood for the initial parameter values is -138.29892658976
#> Optimizing the likelihood - this may take a while.
#> 0.0565371031208431 0.0565371031208431 4 -138.29892658976
#> -19.6874997974797 0.0565371031208431 4 -Inf
#> 0.0565371031208431 -19.6874997974797 4 -Inf
#> 0.0565371031208431 0.0565371031208431 -2.25 -Inf
#> 2.57370518668921 2.57370518668921 -0.166666666666667 -Inf
#> 1.23970037744085 1.23970037744085 0.428571428571429 -Inf
#> 1.23970037744085 0.0565371031208431 4 -Inf
#> 0.0565371031208431 1.23970037744085 4 -408.954085121498
#> 0.0565371031208431 0.0565371031208431 -4.33333333333333 -Inf
#> 0.630909092451593 0.630909092451593 0.428571428571429 -394.970765338453
#> -0.183184094316379 1.55798479466909 0.875 -Inf
#> -0.0289580508740217 0.887427671713442 1.22222222222222 -Inf
#> 0.435774310919359 0.0565371031208431 4 -217.435326543504
#> 0.0565371031208431 0.435774310919359 4 -215.759921685576
#> 0.282344532762055 0.282344532762055 1.22222222222222 -235.408077326566
#> 0.0565371031208431 0.056537103120843 -21 -Inf
#> 0.217302799841525 0.217302799841525 2.07692307692308 -202.86182250914
#> -0.164220824193095 0.0565371031208431 4 -Inf
#> 0.0565371031208431 -0.164220824193095 4 -Inf
#> 0.0565371031208431 0.0565371031208431 1.22222222222222 -140.359130113342
#> -0.101652478250067 0.435774310919359 1.72727272727273 -Inf
#> -0.0667186885311693 0.217302799841525 2.07692307692308 -Inf
#> -0.0667186885311693 0.0565371031208431 4 -Inf
#> 0.0565371031208431 -0.0667186885311693 4 -Inf
#> 0.0565371031208431 0.0565371031208431 2.07692307692308 -139.142155325368
#> -0.0289580508740217 0.217302799841525 2.52941176470588 -Inf
#> -0.00890822400049493 0.13123669972533 2.80952380952381 -Inf
#> -0.00890822400049493 0.0565371031208431 4 -Inf
#> 0.0565371031208431 -0.00890822400049493 4 -Inf
#> 0.0565371031208431 0.0565371031208431 2.80952380952381 -138.67921314246
#> 0.13123669972533 0.0565371031208431 4 -147.961630608301
#> 0.0565371031208431 0.13123669972533 4 -148.307993814251
#> 0.0565371031208431 0.0565371031208431 6.27272727272728 -138.007331731512
#> 0.105190205544315 -0.00890822400049493 5.31578947368422 -Inf
#> 0.0682943175325757 0.0926116256780313 4.27472527472528 -143.832954528161
#> -0.00203968049049697 0.0803161467061682 5.60550458715597 -Inf
#> 0.0946881343274375 0.0623831825484161 4.32347504621072 -143.090390803217
#> 0.0696168471619277 0.0264136556123879 5.22478386167147 -136.921872814753
#> 0.0702793406885307 -0.00376572610556988 5.84085510688836 -Inf
#> 0.0290624149289556 0.0306992264027016 5.93233484889008 -138.621348872489
#> 0.0447199562294352 0.0384417109814988 5.44535621036927 -136.933297942811
#> 0.0571834882945641 0.0245878726542851 8.78482446206113 -136.628875452395
#> 0.0575069775640771 0.00932706313748703 17.7622149837132 -137.320525917318
#> 0.0576148513184024 0.00434065127435459 6.08390270565726 -137.660662938957
#> 0.0573452081868561 0.01690021567186 6.13018361873322 -136.854746501097
#> 0.0785180903673741 0.00726785049012128 7.80434782608691 -136.66692497088
#> 0.0589471061052005 0.00619256503065626 12.0371363903092 -137.452222887806
#> 0.0669293051341006 0.0212825683726463 6.16013858035703 -136.672632060038
#> 0.0777955629677048 0.0184151658513158 9.33996720003544 -136.875767566307
#> 0.0623846944917256 0.0172785305143808 6.73008650701768 -136.669341578847
#> 0.0649759220488105 0.0114221191842328 10.0578020753919 -136.754385324718
#> 0.0664402878763672 0.0187994754408668 6.85206414499345 -136.62941971293
#> 0.0722796266989215 0.0163897753025553 9.06121371968436 -136.664259288779
#> 0.0523333066096938 0.0328819777396786 8.45922624799519 -136.644991100937
#> 0.0571834882945641 0.0245878726542851 10.5508021390374 -136.61329407849
#> 0.0571834882945641 0.00844351337212945 8.78482446206113 -137.355370383151
#> 0.0749398740088071 0.0245878726542851 8.78482446206113 -137.000930185751
#> 0.068955173283337 0.0412575580652741 9.89533417402267 -137.62465214392
#> 0.060102036129208 0.0164515916081767 9.04067402672863 -136.731129320048
#> 0.0418853769445456 0.0191492893092306 10.1053984575835 -137.888330604895
#> 0.0664811703227776 0.0232227923625168 9.08462211847093 -136.661966791511
#> 0.0604273142411855 0.0319302561674392 9.82706766917288 -136.710722399657
#> 0.0603459760007705 0.0280165617887137 9.61914272545702 -136.632090775921
#> 0.0501164607402297 0.0282459571605786 10.1772315653298 -136.685285960957
#> 0.0623423617057083 0.0244739768629935 9.33724669109395 -136.618106435315
#> 0.0574530489386094 0.0211065491733346 9.39815464848055 -136.67481972123
#> 0.0596212613135387 0.0262803056984378 9.56301975217448 -136.609547399912
#> 0.0622516599521428 0.0256394398772377 11.0305172263747 -136.618678751559
#> 0.0609800668078614 0.0253763457627227 10.3776997414283 -136.606982053466
#> 0.0561940323902956 0.0263565014061938 11.0939591443943 -136.601916434086
#> 0.0531464784189513 0.0273003604917051 12.2170088677485 -136.620072375956
#> 0.0561940323902956 0.0263565014061938 11.6936320646434 -136.598657837232
#> 0.0561940323902956 0.0304879388893961 11.0939591443943 -136.601797973284
#> 0.0518543360127933 0.0263565014061938 11.0939591443943 -136.67522313991
#> 0.0605696861929951 0.0291071023570987 11.4872408653628 -136.638485430601
#> 0.0583773367400651 0.0284180708851324 11.3865418216029 -136.601123650355
#> 0.0576485663801091 0.0304879388893961 11.6936320646434 -136.611870687239
#> 0.0565572906928047 0.0273862519522353 11.2385020583417 -136.593935947306
#> 0.0578913782940932 0.0243031806466864 11.7994074348409 -136.603654137303
#> 0.056617858037715 0.0289347579451092 11.2629291226824 -136.594303984702
#> 0.0545423449410474 0.0266995221514903 11.4032120586783 -136.613573793234
#> 0.0574159761990854 0.0279878945637536 11.3907051784793 -136.593337698077
#> 0.0575339439395769 0.029854595782983 10.9244371603553 -136.609229071358
#> 0.0565286918602074 0.0272287952443322 11.4921777044809 -136.592899600194
#> 0.057049882100307 0.0261374692352738 11.4848913607935 -136.596049082943
#> 0.05672583094284 0.0282340069577987 11.3176764761025 -136.592637544749
# ML -136.5796